Tidy up the AudioChart CSUN etc. presentation
* Update the packaging (simpler).
* Move to impress.js as jmpress.js is not being developed.
* Have only the latest version.
* ACK TPG in text as the logo is no longer there in the latest version.
* This fixes the security warnings and removes or updates outdated packages.

Note: I have been a bit naughty and have been doing some rebasing (all
nicely educational) to clean up the commit history a bit. I managed to
do sufficient submodule-adding-wrangling to get the state of the repo to
be fine after the AudioChart presentation was renamed (or will be
renamed, from our current point of view—wibbly-wobbly timey-wimey :-)),
however I have not gone back to fix the paths in this commit (only to
have to fix them again later), partly due to one of the virtues of
programmers, partly for pragmatic reasons (other things to be doing :-))
and partly so I can keep this note here to remind myself to not do this
sort of thing willy-nilly :-).